Review feedback
Many changes are not intended to be committed as is. Feel free to
take some of them, but some are just review comments.

Added a .github with the publish action.
If you make a release on GitHub it will automatically publish the
package on pkg.toit.io.

Added package.yaml/lock. This is necessary for the package to work on
pkg.toit.io.

renamed the driver.toit to cs5529.toit. No need to have this
redirection. (Let me know where you copied it from, so I can fix it
    there).

Minor changes in the source code.